pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pbi-developer"
version = "0.1.0"
description = "AI-powered Power BI developer tool using Claude Sonnet"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.40.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.0",
    "python-pptx>=1.0",
    "opencv-python>=4.9",
    "Pillow>=10.0",
    "msal>=1.28",
    "requests>=2.31",
    "jsonschema>=4.21",
    "snowflake-connector-python>=3.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
deploy = [
    "fabric-cicd>=0.2",
]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
pbi-dev = "pbi_developer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/pbi_developer/cli.py" = ["B008"]  # typer.Option/Argument in defaults is standard Typer pattern

[tool.ruff.lint.isort]
known-first-party = ["pbi_developer"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
